backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
mtcnn==0.1.1
keras-facenet==0.3.2
scipy==1.11.4
pyttsx3==2.90
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
